+++ lib/utils/config_file/toolbar_config.dart
@@ -1,1 +1,0 @@
-// TODO: Add toolbar configurations to .printnotes/main_config.json